fix #1896, move storm examples under integration test
