+++ integrationtest/core/src/it/resources/log4j.properties
