+++ integrationtest/core/src/it/scala/io/gearpump/integrationtest/checklist/CommandLineSpec.scala
