+++ integrationtest/core/src/main/scala/io/gearpump/integrationtest/Docker.scala
