+++ integrationtest/core/src/main/scala/io/gearpump/integrationtest/minicluster/CommandLineClient.scala
