+++ integrationtest/core/src/main/scala/io/gearpump/integrationtest/minicluster/RestClient.scala
